Removed MainController
