+++ src/main/java/address/controller/PersonListPanel.java
@@ -2,6 +2,7 @@ package address.controller;
 
 import address.model.ModelManager;
 import address.model.datatypes.person.ReadOnlyViewablePerson;
+import address.ui.Ui;
 import address.util.collections.UnmodifiableObservableList;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
@@ -14,11 +15,11 @@ public class PersonListPanel extends BaseUiPart {
     private PersonListPanelView view;
     private PersonListPanelController controller;
 
-    public PersonListPanel(Stage primaryStage, AnchorPane pane, MainController mainController,
+    public PersonListPanel(Stage primaryStage, AnchorPane pane, Ui ui,
                            ModelManager modelManager, UnmodifiableObservableList<ReadOnlyViewablePerson> personList) {
         super(primaryStage);
         view = new PersonListPanelView(primaryStage, pane);
         controller = view.getLoader().getController();
-        controller.setConnections(mainController, modelManager, personList);
+        controller.setConnections(ui, modelManager, personList);
     }
 }
