+++ src/main/java/address/controller/StatusBarHeaderController.java
@@ -2,6 +2,7 @@ package address.controller;
 
 
 import address.events.SingleTargetCommandResultEvent;
+import address.ui.Ui;
 import address.util.CommandResultFormatter;
 import javafx.application.Platform;
 import javafx.collections.ListChangeListener;
@@ -11,17 +12,17 @@ import org.controlsfx.control.StatusBar;
 public class StatusBarHeaderController extends UiController{
 
     private StatusBar headerStatusBar;
-    private MainController mainController;
+    private Ui ui;
     private ObservableList<SingleTargetCommandResultEvent> finishedCommands;
 
-    public StatusBarHeaderController(MainController mainController, ObservableList<SingleTargetCommandResultEvent> finishedCommands) {
-        this.mainController = mainController;
+    public StatusBarHeaderController(Ui ui, ObservableList<SingleTargetCommandResultEvent> finishedCommands) {
+        this.ui = ui;
         this.finishedCommands = finishedCommands;
         headerStatusBar = new StatusBar();
         headerStatusBar.getStyleClass().removeAll();
         headerStatusBar.getStyleClass().add("status-bar-with-border");
         headerStatusBar.setText("");
-        headerStatusBar.setOnMouseClicked(event -> mainController.showActivityHistoryDialog());
+        headerStatusBar.setOnMouseClicked(event -> ui.showActivityHistoryDialog());
         this.finishedCommands.addListener((ListChangeListener<SingleTargetCommandResultEvent>) c -> {
             SingleTargetCommandResultEvent lastCommandInfo = finishedCommands.get(finishedCommands.size() - 1);
             Platform.runLater(() -> headerStatusBar.setText(CommandResultFormatter.getStringRepresentation(lastCommandInfo)));
