+++ src/main/java/address/controller/TagListController.java
@@ -2,6 +2,7 @@ package address.controller;
 
 import address.model.datatypes.tag.Tag;
 import address.model.ModelManager;
+import address.ui.Ui;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.ScrollPane;
@@ -13,11 +14,11 @@ import javafx.stage.Stage;
 /**
  * Dialog to show the list of available tags
  *
- * Stage, tags, mainController and modelManager should be set before showing stage
+ * Stage, tags, ui and modelManager should be set before showing stage
  */
 public class TagListController extends UiController{
     Stage stage;
-    MainController mainController;
+    Ui ui;
     ModelManager modelManager;
 
     @FXML
@@ -36,27 +37,27 @@ public class TagListController extends UiController{
     }
 
     public void setTags(ObservableList<Tag> tagList) {
-        tags.setContent(getTagsVBox(tagList, mainController));
+        tags.setContent(getTagsVBox(tagList, ui));
     }
 
-    public void setMainController(MainController mainController) {
-        this.mainController = mainController;
+    public void setUi(Ui ui) {
+        this.ui = ui;
     }
 
     public void setModelManager(ModelManager modelManager) {
         this.modelManager = modelManager;
     }
 
-    public VBox getTagsVBox(ObservableList<Tag> tagList, MainController mainController) {
+    public VBox getTagsVBox(ObservableList<Tag> tagList, Ui ui) {
         VBox vBox = new VBox();
 
         if (tagList.size() == 0) {
-            vBox.getChildren().add(TagCardController.getDummyTagCard(this, mainController));
+            vBox.getChildren().add(TagCardController.getDummyTagCard(this, ui));
             return vBox;
         }
         tagList.stream()
                 .forEach(tag -> {
-                    TagCardController tagCardController = new TagCardController(tag, mainController, this);
+                    TagCardController tagCardController = new TagCardController(tag, ui, this);
                     vBox.getChildren().add(tagCardController.getLayout());
                 });
         return vBox;
